Moves OperatorValidator to operators_validator
