+++ src/lib.rs
@@ -69,11 +69,12 @@ pub use primitives::Type;
 pub use primitives::V128;
 
 pub use validator::validate;
-pub use validator::OperatorValidatorConfig;
 pub use validator::ValidatingOperatorParser;
 pub use validator::ValidatingParser;
 pub use validator::ValidatingParserConfig;
-pub use validator::WasmModuleResources;
+
+pub use operators_validator::OperatorValidatorConfig;
+pub use operators_validator::WasmModuleResources;
 
 pub use readers::CodeSectionReader;
 pub use readers::Data;
@@ -116,6 +117,7 @@ pub use readers::TypeSectionReader;
 
 mod binary_reader;
 mod limits;
+mod operators_validator;
 mod parser;
 mod primitives;
 mod readers;
